fix(BBS-Ruby): fix checking API ver against a wrong TOC item
* "## Version" -> "## Interface"

The "## Version" TOC is the version name given by the user
for the BBS-Ruby script written by the user.
It is irrelevant to the version of the BBS-Ruby interpreter.

The code with this issue was enabled by 7b57676904bcc6c465e605d105b239febcb65fbf (ccns/dreambbs@1ffcbc0960eb7053e8f08cb81dc2cc4cc37389c2).