chore(macros): enclose param in `()` if not surrounded by `()`/`,`
For a macro param which only accepts a C expr containing no non-enclosed `,`s,
the surrounding `()`/`,` ensure that the expanded expr
does not form another expr with the adjacent code in an unexpected way.

For example,
every macro param occurs solely as a param list element to a function/macro
is not an operand of any oprs,
so that the macro param may expand to any C expr without non-enclosed `,`
while not forming an unexpected expr with the adjacent code.

Enclosing such params in `()` in macro definition is generally not necessary,
except when the corresponding arg is possibly a macro
which produces any non-enclosed `,`s after expansion.
In such case, the enclosing `()` should be kept,
especially when such expansion is known to occur.

This commit excludes:
* Params which do not accept a non-enclosed C expr as their argument